Gui formatting tweaks
Prevent the changelog from loading on default. Change the main panel border to an upper border on the ticker. Don't use a margin on the main panel, instead affix the ticker to the bottom of the page, so that the window is not scrollable for a ticker. Overflow on the panels is auto, to remove the unused scrollbar at the bottom of panels.

There are still some formatting issues. Still needs some work to render correctly at different window sizes.