SQL query menu
There is another button that can be uncommented, which adds nested groups. Grouping is why I placed the 'add rule' button on the sql menu instead of the subsetting gui. Having an add button on the subsetting gui is ambiguous-- IE: Which group should a stage/subset be put into?